remove: 필요없는 테스트 및 코드 추가 삭제
