+++ frontend/src/app/_components/add/course/AddCourse.tsx
@@ -3,7 +3,6 @@
 import NaverContext from '@/context/NaverContext';
 import CourseWithNewStepProvider from '@/context/travel/step/course/CourseWithNewStepContext';
 import CourseWithoutNewStepProvider from '@/context/travel/step/course/CourseWithoutNewStepContext';
-import PlaceProvider from '@/context/travel/place/PlaceContext';
 import CourseProvider from '@/context/travel/course/CourseContext';
 import AddTravelMapProvider from '@/context/map/AddTravelMapContext';
 import SearchProvider from '@/context/search/SearchContext';
@@ -19,19 +18,17 @@ export default function AddCourse({
   return (
     <section className="h-full flex flex-col grow">
       <NaverContext ncpClientId={ncpClientId}>
-        <PlaceProvider course>
-          <CourseProvider>
-            <CourseWithNewStepProvider>
-              <CourseWithoutNewStepProvider>
-                <AddTravelMapProvider>
-                  <SearchProvider initialKeyword={initialKeyword}>
-                    <AddCourseContent />
-                  </SearchProvider>
-                </AddTravelMapProvider>
-              </CourseWithoutNewStepProvider>
-            </CourseWithNewStepProvider>
-          </CourseProvider>
-        </PlaceProvider>
+        <CourseProvider>
+          <CourseWithNewStepProvider>
+            <CourseWithoutNewStepProvider>
+              <AddTravelMapProvider>
+                <SearchProvider initialKeyword={initialKeyword}>
+                  <AddCourseContent />
+                </SearchProvider>
+              </AddTravelMapProvider>
+            </CourseWithoutNewStepProvider>
+          </CourseWithNewStepProvider>
+        </CourseProvider>
       </NaverContext>
     </section>
   );
