+++ frontend/src/app/_components/add/spot/AddSpot.tsx
@@ -2,7 +2,6 @@
 
 import NaverContext from '@/context/NaverContext';
 import SpotStepProvider from '@/context/travel/step/spot/SpotStepContext';
-import PlaceProvider from '@/context/travel/place/PlaceContext';
 import SpotProvider from '@/context/travel/spot/SpotContext';
 import AddTravelMapProvider from '@/context/map/AddTravelMapContext';
 import SearchProvider from '@/context/search/SearchContext';
@@ -18,17 +17,15 @@ export default function AddSpot({
   return (
     <section className="h-full flex flex-col grow">
       <NaverContext ncpClientId={ncpClientId}>
-        <PlaceProvider>
-          <SpotProvider>
-            <SpotStepProvider>
-              <AddTravelMapProvider>
-                <SearchProvider initialKeyword={initialKeyword}>
-                  <AddSpotContent />
-                </SearchProvider>
-              </AddTravelMapProvider>
-            </SpotStepProvider>
-          </SpotProvider>
-        </PlaceProvider>
+        <SpotProvider>
+          <SpotStepProvider>
+            <AddTravelMapProvider>
+              <SearchProvider initialKeyword={initialKeyword}>
+                <AddSpotContent />
+              </SearchProvider>
+            </AddTravelMapProvider>
+          </SpotStepProvider>
+        </SpotProvider>
       </NaverContext>
     </section>
   );
